requirements.txt
SQLAlchemy==2.0.*
Flask-SQLAlchemy==3.1.*
python-dotenv==1.*
Flask-WTF==1.*
Flask==3.0.*
